samples/hosted-agent/python/fileSearchAgent/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b16
debugpy
# Agent development CLI tool (preview)
agent-dev-cli